+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added error message when accessing `ReactCommands` before `ReactPlugin` has been added.
 - Added `RawCallbackSystem` as a system-runner helper that doesn't require boxing the system. Used this to optimize system command spawning.
 - Added perf warning for when a large number of reactors are registered that are triggered by a specific entity.
+- Added infinite loop detection to reaction trees.
 
 
 ## [0.10.0]